Modularized APIs to independent controllers, now we can test the API
