+++ stegonoserver/app/tests/conftest.py
@@ -6,8 +6,10 @@ import time
 import pytest
 from PIL import Image
 
+
 parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.append(parent_dir)
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
 
 
 # Sample abstract images used for testing:
